chore(llc): align Tezos signer in `generic-alpaca`
