+++ .changeset/itchy-tables-look.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/coin-tezos": patch
+"@ledgerhq/live-common": patch
+---
+
+chore(llc): align Tezos signer in `generic-alpaca`
