+++ libs/ledger-live-common/src/families/tezos/getAddress.ts
@@ -1,7 +1,7 @@
 import { GetAddressFn } from "@ledgerhq/ledger-wallet-framework/bridge/getAddressWrapper";
 import { GetAddressOptions } from "@ledgerhq/ledger-wallet-framework/derivation";
 import { SignerContext } from "@ledgerhq/ledger-wallet-framework/signer";
-import type { TezosSigner } from "../types";
+import type { TezosSigner } from "./types";
 
 const getAddress = (signerContext: SignerContext<TezosSigner>): GetAddressFn => {
   return async (deviceId: string, { path, verify }: GetAddressOptions) => {
